+++ resources/config/taggable.php
@@ -60,4 +60,14 @@ return [
      * then update the configuration below.
      */
     'model'  => \Cviebrock\EloquentTaggable\Models\Tag::class,
+
+
+    /**
+     * The tables used to store the tags in the database.  You can
+     * publish the package's migrations and use custom names.
+     */
+    'tables' => [
+        'taggable_tags'      => 'taggable_tags',
+        'taggable_taggables' => 'taggable_taggables',
+    ]
 ];