+++ src/Models/Tag.php
@@ -16,7 +16,7 @@ class Tag extends Model
     /**
      * @inheritdoc
      */
-    protected $table = 'taggable_tags';
+    protected $table;
 
     /**
      * @inheritdoc
@@ -39,6 +39,9 @@ class Tag extends Model
         if ($connection = config('taggable.connection')) {
             $this->setConnection($connection);
         }
+
+        $table = config('taggable.tables.taggable_tags', 'taggable_tags');
+        $this->setTable($table);
 
         parent::__construct($attributes);
     }
@@ -102,7 +105,9 @@ class Tag extends Model
      */
     protected function taggedModels(string $class): MorphToMany
     {
-        return $this->morphedByMany($class, 'taggable', 'taggable_taggables', 'tag_id');
+        $table = config('taggable.tables.taggable_taggables', 'taggable_taggables');
+
+        return $this->morphedByMany($class, 'taggable', $table, 'tag_id');
     }
 
     /**