+++ src/ServiceProvider.php
@@ -20,11 +20,17 @@ class ServiceProvider extends LaravelServiceProvider
     /**
      * Bootstrap the application events.
      */
-    public function boot()
+    public function boot(): void
     {
         $this->publishes([
             __DIR__ . '/../resources/config/taggable.php' => config_path('taggable.php'),
         ], 'config');
+
+        if (!class_exists('CreateTaggableTable')) {
+            $this->publishes([
+                __DIR__.'/../resources/database/migrations/create_taggable_table.php.stub' => database_path('migrations/' . date('Y_m_d_His') . '_create_taggable_table.php'),
+            ], 'migrations');
+        }
 
         $this->loadMigrationsFrom(
             __DIR__.'/../resources/database/migrations'