+++ tests/TestCase.php
@@ -17,10 +17,11 @@ abstract class TestCase extends Orchestra
     {
         parent::setUp();
 
-        $this->artisan('migrate', ['--database' => 'test']);
+        $this->setUpDatabase();
 
-        $this->beforeApplicationDestroyed(function() {
-            $this->artisan('migrate:rollback');
+        $this->beforeApplicationDestroyed(static function() {
+            (new \CreateTestModelsTable())->down();
+            (new \CreateTaggableTable())->down();
         });
     }
 
@@ -90,5 +91,17 @@ abstract class TestCase extends Orchestra
     protected function newDummy(array $data = ['title' => 'dummy']): TestDummy
     {
         return TestDummy::create($data);
+    }
+
+    /**
+     * Set up the database
+     */
+    private function setUpDatabase(): void
+    {
+        include_once __DIR__.'/../resources/database/migrations/create_taggable_table.php.stub';
+        (new \CreateTaggableTable())->up();
+
+        include_once __DIR__.'/database/migrations/2013_11_04_163552_create_test_models_table.php';
+        (new \CreateTestModelsTable())->up();
     }
 }